Move SendmailClient to a separate file
